Added a flag to the compiler function(s) deciding whether to compile .clj files as well.
The logic right now is as follows:

Both .cljs and clj files are tried, both read into the JVM and to be compiled
to JavaScript. BUT, if a file ends in .clj it has to have a speical
marker form somewhere in the source in order to result in a JS file,
and you will get a warning message if it removed the generated JS file.

The special marker is

    (comment *target* clojurescript)

So, now you can use regular :require clauses in your ns declaration,
and also use .clj as suffix for all files in the project, no matter
whether it is on server side, client side or macros for client side
purpose.