The compiler feature/directive system now works.
One can specify directives in the form of

   (comment **compiler** :feature value)

Compiler features are also heuristically assessed by the system.
Heuristics is a big word for checking that the filename extension
is ".cljs"...

With this system in place, one can now safely use .clj for
ClojureScript files and place macros in source files used by both
the macro system (in JVM) and the target JavaScript code, it is
just so that the macros will be ignored in runtime.